Add client tests
