+++ .github/workflows/test.yml
@@ -35,3 +35,16 @@ jobs:
         run: SOURCING=1 . run && ${{ matrix.test-args }} build_client
       - name: Compile and run tests
         run: TESTS=server ${{ matrix.test-args }} ./test
+  test-client:
+    name: Test client
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        test-args:
+          - ""
+          - RELEASE=1
+      fail-fast: false
+    steps:
+      - uses: actions/checkout@v2
+      - name: Compile and run tests
+        run: TESTS=client ${{ matrix.test-args }} ./test
