+++ client/jest.config.js
@@ -1,0 +1,20 @@
+module.exports = {
+  // The root of your source code, typically /src
+  // `<rootDir>` is a token Jest substitutes
+  roots: ["<rootDir>/src"],
+
+  // Jest transformations -- this adds support for TypeScript using ts-jest
+  transform: {
+    "^.+\\.tsx?$": "ts-jest",
+  },
+
+  // Test spec file resolution pattern
+  // Filename should contain `test` or `spec`
+  testRegex: "test.tsx?$",
+
+  // Module file extensions for importing
+  moduleFileExtensions: ["ts", "tsx", "js", "jsx", "json", "node"],
+
+  // https://jestjs.io/docs/configuration#testenvironment-string
+  testEnvironment: "jsdom",
+};
