+++ client/src/App.tsx
@@ -1,15 +1,11 @@
 import * as React from "react";
-import * as ReactDOM from "react-dom";
-
-declare var __NAME__: string;
-declare var __VERSION__: string;
 
 interface AppComponentProps {
   name: string;
   version: string;
 }
 
-class App extends React.Component<AppComponentProps> {
+export default class App extends React.Component<AppComponentProps> {
   render() {
     return (
       <p>
@@ -18,10 +14,3 @@ class App extends React.Component<AppComponentProps> {
     );
   }
 }
-
-const rootElement = document.getElementById("root") as HTMLElement;
-ReactDOM.render(<App name={__NAME__} version={__VERSION__} />, rootElement);
-
-/// #if DEBUG
-console.log(`${__NAME__} v${__VERSION__}`);
-/// #endif
