+++ client/tsconfig.json
@@ -1,7 +1,8 @@
 {
   "compilerOptions": {
     "sourceMap": true,
-    "jsx": "react"
+    "jsx": "react",
+    "esModuleInterop": true
   },
   "include": ["./src/**/*.tsx"]
 }
